Remove mc binary. Impl tests
