+++ algorithms/src/MarchingCubesConfig.h
@@ -14,7 +14,7 @@ namespace SPHAlgorithms
 constexpr float CUBE_SIZE = 3.0f;
 
 // Number of the grid cubes by which the domain is splitted
-constexpr int GRID_CUBES_NUMBER = 300;
+constexpr int GRID_CUBES_NUMBER = 100;
 
 // Size of the grid cube
 const constexpr float GRID_CUBE_SIZE = CUBE_SIZE / GRID_CUBES_NUMBER;
