+++ algorithms/test/src/MarchingCubesTestSuite.h
@@ -1,0 +1,26 @@
+/**
+ * @file MarchingCubesTestSuite.h
+ * @author Anton Artyukh (artyukhanton@gmail.com)
+ * @date Created May 27, 2019
+ * @par Anton Artiukh
+ **/
+
+#ifndef MARCHING_CUBES_TEST_SUITE_H_96192C2023784EE0B4976A48A1A8779R
+#define MARCHING_CUBES_TEST_SUITE_H_96192C2023784EE0B4976A48A1A8779R
+
+namespace SPHAlgorithms
+{
+
+namespace TestEnvironment
+{
+
+class MarchingCubesTestSuite
+{
+public:
+    static void generatePawnMesh();
+};
+
+} // namespace TestEnvironment
+} // namespace SPHAlgorithms
+
+#endif // MARCHING_CUBES_TEST_SUITE_H_96192C2023784EE0B4976A48A1A8779R
